Split up licenses and shared model to appropriate test and files
