+++ exasol/toolbox/nox/_dependencies.py
@@ -3,189 +3,19 @@ from __future__ import annotations
 import argparse
 import json
 import subprocess
-import tempfile
-from inspect import cleandoc
-from json import loads
 from pathlib import Path
 
 import nox
 from nox import Session
 
+from exasol.toolbox.util.dependencies.licenses import (
+    licenses,
+    packages_to_markdown,
+)
 from exasol.toolbox.util.dependencies.poetry_dependencies import (
-    Package,
     PoetryDependencies,
-    PoetryDependency,
     PoetryToml,
 )
-
-
-class PackageLicense(Package):
-    package_link: str
-    license: str
-    license_link: str
-
-
-def _normalize(_license: str) -> str:
-    def is_multi_license(l):
-        return ";" in l
-
-    def select_most_restrictive(licenses: list) -> str:
-        _max = 0
-        lic = "Unknown"
-        _mapping = {
-            "Unknown": -1,
-            "Unlicensed": 0,
-            "BSD": 1,
-            "MIT": 2,
-            "MPLv2": 3,
-            "LGPLv2": 4,
-            "GPLv2": 5,
-            "GPLv3": 6,
-        }
-        for l in licenses:
-            if l in _mapping:
-                if _mapping[l] > _mapping[lic]:
-                    lic = l
-            else:
-                return "<br>".join(licenses)
-        return lic
-
-    mapping = {
-        "BSD License": "BSD",
-        "MIT License": "MIT",
-        "The Unlicensed (Unlicensed)": "Unlicensed",
-        "Mozilla Public License 2.0 (MPL 2.0)": "MPLv2",
-        "GNU General Public License (GPL)": "GPL",
-        "GNU Lesser General Public License v2 (LGPLv2)": "LGPLv2",
-        "GNU General Public License v2 (GPLv2)": "GPLv2",
-        "GNU General Public License v2 or later (GPLv2+)": "GPLv2+",
-        "GNU General Public License v3 (GPLv3)": "GPLv3",
-        "Apache Software License": "Apache",
-    }
-
-    if is_multi_license(_license):
-        items = []
-        for item in _license.split(";"):
-            item = str(item).strip()
-            if item in mapping:
-                items.append(mapping[item])
-            else:
-                items.append(item)
-        return select_most_restrictive(items)
-
-    if _license not in mapping:
-        return _license
-
-    return mapping[_license]
-
-
-def _packages_from_json(json: str) -> list[PackageLicense]:
-    packages = loads(json)
-    packages_list = []
-    mapping = {
-        "GPLv1": "https://www.gnu.org/licenses/old-licenses/gpl-1.0.html",
-        "GPLv2": "https://www.gnu.org/licenses/old-licenses/gpl-2.0.html",
-        "LGPLv2": "https://www.gnu.org/licenses/old-licenses/lgpl-2.0.html",
-        "GPLv3": "https://www.gnu.org/licenses/gpl-3.0.html",
-        "LGPLv3": "https://www.gnu.org/licenses/lgpl-3.0.html",
-        "Apache": "https://www.apache.org/licenses/LICENSE-2.0",
-        "MIT": "https://mit-license.org/",
-        "BSD": "https://opensource.org/license/bsd-3-clause",
-    }
-    for package in packages:
-        package_license = _normalize(package["License"])
-        packages_list.append(
-            PackageLicense(
-                name=package["Name"],
-                package_link="" if package["URL"] == "UNKNOWN" else package["URL"],
-                version=package["Version"],
-                license=package_license,
-                license_link=(
-                    "" if package_license not in mapping else mapping[package_license]
-                ),
-            )
-        )
-    return packages_list
-
-
-def _licenses() -> list[PackageLicense]:
-    with tempfile.NamedTemporaryFile() as file:
-        subprocess.run(
-            [
-                "poetry",
-                "run",
-                "pip-licenses",
-                "--format=json",
-                "--output-file=" + file.name,
-                "--with-system",
-                "--with-urls",
-            ],
-            capture_output=True,
-        )
-        result = _packages_from_json(file.read().decode())
-    return result
-
-
-def _packages_to_markdown(
-    dependencies: dict[str, list], packages: list[PackageLicense]
-) -> str:
-    def heading():
-        text = "# Dependencies\n"
-        return text
-
-    def dependency(
-        group: str,
-        group_packages: list[PoetryDependency],
-        packages: list[PackageLicense],
-    ) -> str:
-        def _header(_group: str):
-            _group = "".join([word.capitalize() for word in _group.strip().split()])
-            text = f"## {_group} Dependencies\n"
-            text += "|Package|version|Licence|\n"
-            text += "|---|---|---|\n"
-            return text
-
-        def _rows(
-            _group_packages: list[PoetryDependency], _packages: list[PackageLicense]
-        ) -> str:
-            text = ""
-            for package in _group_packages:
-                consistent = filter(
-                    lambda elem: elem.normalized_name == package.normalized_name,
-                    _packages,
-                )
-                for content in consistent:
-                    if content.package_link:
-                        text += f"|[{content.name}]({content.package_link})"
-                    else:
-                        text += f"|{content.name}"
-                    text += f"|{content.version}"
-                    if content.license_link:
-                        text += f"|[{content.license}]({content.license_link})|\n"
-                    else:
-                        text += f"|{content.license}|\n"
-            text += "\n"
-            return text
-
-        _template = cleandoc(
-            """
-            {header}{rows}
-        """
-        )
-        return _template.format(
-            header=_header(group), rows=_rows(group_packages, packages)
-        )
-
-    template = cleandoc(
-        """
-        {heading}{rows}
-    """
-    )
-
-    rows = ""
-    for group in dependencies:
-        rows += dependency(group, dependencies[group], packages)
-    return template.format(heading=heading(), rows=rows)
 
 
 class Audit:
@@ -262,8 +92,8 @@ def dependency_licenses(session: Session) -> None:
     dependencies = PoetryDependencies(
         groups=groups, working_directory=working_directory
     ).direct_dependencies
-    package_infos = _licenses()
-    print(_packages_to_markdown(dependencies=dependencies, packages=package_infos))
+    package_infos = licenses()
+    print(packages_to_markdown(dependencies=dependencies, packages=package_infos))
 
 
 @nox.session(name="dependency:audit", python=False)
